backend/requirements.txt
asgiref==3.11.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.10
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.3
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.2.3
